# Link bitcode device libraries before the optimization pipeline