@e-kayrakli 2023-08-01
> This was a very large PR and so very difficult to review; however, it all looks good to me. I do suspect that a full GPU test is the best way to make sure moving the bitcode linking earlier is safe.

Haha, I was inclined to lump this into the other PR, but they look so irrelevant.

> For my own curiosity: what part of that step do you think performs the optimizations we weren't opting into?

Are you asking how it is done in the compiler? If so:

We generate our compilation pipeline and run it here: https://github.com/chapel-lang/chapel/blob/main/compiler/llvm/clangUtil.cpp#L4725-L4728

It looks like the `PMBuilder` is where the magic is created. We configure it here: https://github.com/chapel-lang/chapel/blob/main/compiler/llvm/clangUtil.cpp#L2285

The call that this PR moves was happening after `mpm.run` that's highlighted in the first link. This PR moves it to before. The key optimization that we were missing is inlining.

I think there's still something that we can improve here. In the generated PTX, I can still see the `__nv_sqrt` definition even though all calls to it are inlined. That function must have external linkage which prevents it from dead-code eliminated. But leaving a dead function in the generated PTX is not something we should worry about at the moment.