@milthorpe 2023-08-03
With this fix, the Chapel compiler inlines the fast math operation as expected for the example code in #22112. I noticed that the generated PTX still includes a wrapper function `__nv_sqrtf` which is not actually called from anywhere. This doesn't seem to be a problem, as although it does increase the size of the fatbinary file `chpl__gpu.fatbin`, it is removed in the linker step so it doesn't actually increase the size of the executable. I'm just noting it here as it is a bit of an anomaly compared to the nvcc-generated PTX from the CUDA code, which doesn't include this wrapper function.

```
    // .globl   __nv_sqrtf
.visible .func  (.param .b32 func_retval0) __nv_sqrtf(
    .param .b32 __nv_sqrtf_param_0
)
{
    .reg .f32   %f<3>;

    ld.param.f32    %f1, [__nv_sqrtf_param_0];
    sqrt.approx.f32     %f2, %f1;
    st.param.f32    [func_retval0+0], %f2;
    ret;

}
```