@e-kayrakli 2023-08-07
@milthorpe -- thanks for noting that. I realized that, but I agree that it is not a problem in the short term.

My theory is that that function has external linkage as it comes from a library. LLVM's inliner has to keep it in, because it thinks it could be callable from a different module.

Somewhere in the optimization pipeline, we could make everything that's not a kernel we generate internal and that should make that function easy to remove for LLVM. It is not obvious to me what's the right way to do that right now. Before diving in writing code for it, I'd probably read some clang code to see how it handles this.