@bradcray 2023-08-07
Hi Josh @milthorpe : Have you had a chance to see where this PR brings your performance relative to the (I think it was) CUDA version you were comparing against?  I'm still haunted by the original ratio you mentioned to us (though I think there turned out to be extenuating circumstances), so am curious how far off we are at this point if you happen to know.  Thanks!