@milthorpe 2023-08-07
Some iteration times for [miniBUDE](https://github.com/xianghao-wang/miniBUDE/tree/benchmark) (large problem size "bm2"), before and after Engin's fix:
| GPU arch | Chapel before PR (s) | Chapel after PR (s) | CUDA (s) |
|---|---:|---:|---:|
| P100 | 11.3 | 7.9 | 4.0 |
| V100 | 5.1 | 2.8 | 2.0 |

We believe that much of the remaining performance gap with CUDA can be explained by lower occupancy due to higher register pressure, as the Chapel-generated kernel has many more parameters and variables than the equivalent CUDA kernel.